+++ src/api/index.tsx
@@ -1,4 +1,11 @@
 export { default as apiClient } from './apiClient';
 export { login, logout } from './services/auth/user/UserSlice';
 export { AuthGoogle } from './services/auth/google/AuthGoogle';
+export { AuthNaver } from './services/auth/naver/AuthNaver';
 export { default as userReducer } from './services/auth/user/UserSlice';
+
+export type {
+    User,
+    SocialState,
+    LoginPayload,
+} from './services/auth/types/types';
