+++ src/components/layout/MainLayout.tsx
@@ -1,31 +1,31 @@
-import { Routes, Route } from "react-router-dom";
-import Header from "./Header";
-import Sidebar from "./Sidebar";
-import MapArea from "../map/MapArea";
-import Community from "../pages/Community";
-import PopularReviewers from "../pages/PopularReviewers";
+import { Routes, Route } from 'react-router-dom';
+import Header from './Header';
+import Sidebar from './Sidebar';
+import MapArea from '../map/MapArea';
+import Community from '../pages/Community';
+import PopularReviewers from '../pages/popularReviewers/PopularReviewersPage';
 
 const MainLayout = () => {
-  return (
-    <div className="h-screen flex flex-col">
-      <Header />
-      <div className="flex flex-1 overflow-hidden">
-        <Routes>
-          <Route
-            path="/"
-            element={
-              <>
-                <Sidebar />
-                <MapArea />
-              </>
-            }
-          />
-          <Route path="/community" element={<Community />} />
-          <Route path="/reviewers" element={<PopularReviewers />} />
-        </Routes>
-      </div>
-    </div>
-  );
+    return (
+        <div className="h-screen flex flex-col">
+            <Header />
+            <div className="flex flex-1 overflow-hidden">
+                <Routes>
+                    <Route
+                        path="/"
+                        element={
+                            <>
+                                <Sidebar />
+                                <MapArea />
+                            </>
+                        }
+                    />
+                    <Route path="/community" element={<Community />} />
+                    <Route path="/reviewers" element={<PopularReviewers />} />
+                </Routes>
+            </div>
+        </div>
+    );
 };
 
 export default MainLayout;
