+++ src/components/pages/index.tsx
@@ -1,0 +1,6 @@
+export { default as AuthNaverCallback } from './callback/AuthNaverCallback';
+export { default as CommunityPage } from './community/CommunityPage';
+export { default as MainPage } from './main/MainPage';
+export { default as PopularReviewersPage } from './popularReviewers/PopularReviewersPage';
+export { default as RecommendedRestaurantsPage } from './recommendedRestaurants/RecommendedRestaurantsPage';
+export { default as RestaurantDetailPage } from './restaurantDetail/RestaurantDetailPage';
